lib: spawn multiple tasks when closing PRs
We can end up wanting to delete quite a few references, so doing it
concurrently can speed things up quite a bit.

It looks like there's a 10s timeout in GitHub when waiting for a webhook
to respond.  Without this, we were hitting that at around 25 iterations
on a PR (50 references).