readme: github: change reference namespace
Due to the fact that there's a limited number of namespaces that work
with Github's GraphQL [1], we sadly need to put references under
refs/heads/.  We need GraphQL when when closing a PR so that we can
delete a large number of references with a single API call.  That change
is forthcoming.

1. https://github.com/orgs/community/discussions/83980