lib: handle closing PRs as a background task
We can end up with a lot of references to remove when closing a PR.  We
don't want to do this with a mass number of individual API requests
because that could blow through our quota.  However, when using GraphQL,
it still takes over 10s to delete just 50 references.  Given that, we
have no real choice but to run this task in the background and report
success to GitHub before it decides to hang up on us.